[ALPS02959686] cmdq: Refine stress error log
1. Add timeout flag log in start of stress test.
2. Reset error number when reset first dump.
3. Remove dump invalid thread by engine flag.
4. Refine code for 32/64 bit pointer compatible.

MTK-Commit-Id: d8b141d56ed73114e978da77d8ffb63c788eb73f

CR-Id: ALPS02959686
Feature: CMDQ

+++ drivers/misc/mediatek/cmdq/v3/cmdq_driver.c
@@ -504,7 +504,7 @@ static long cmdq_ioctl(struct file *pFile, unsigned int code, unsigned long para
 
 		/* insert private_data for resource reclaim */
 		desc_private.node_private_data = pFile->private_data;
-		command.privateData = (cmdqU32Ptr_t)&desc_private;
+		command.privateData = (cmdqU32Ptr_t)(unsigned long)&desc_private;
 
 		if (cmdq_driver_process_command_request(&command))
 			return -EFAULT;
@@ -527,7 +527,7 @@ static long cmdq_ioctl(struct file *pFile, unsigned int code, unsigned long para
 
 		/* insert private_data for resource reclaim */
 		desc_private.node_private_data = pFile->private_data;
-		job.command.privateData = (cmdqU32Ptr_t)&desc_private;
+		job.command.privateData = (cmdqU32Ptr_t)(unsigned long)&desc_private;
 
 		/* create kernel-space address buffer */
 		status = cmdq_driver_create_reg_address_buffer(&job.command);
